# Refactors json codec to reduce redundant list creation