--- a/benchmarks/src/main/resources/span-client.json
+++ b/benchmarks/src/main/resources/span-client.json
@@ -1,48 +1,79 @@
 {
-  "traceId": "5af7183fb1d4cf5f",
-  "name": "query",
-  "id": "352bff9a74ca9ad2",
-  "parentId": "6b221d5bc9e6496c",
-  "timestamp": 1461750040359000,
-  "duration": 5000,
+  "traceId": "86154a4ba6e91385",
+  "name": "get",
+  "id": "4d1e00c0db9010db",
+  "parentId": "86154a4ba6e91385",
+  "timestamp": 1472470996199000,
+  "duration": 207000,
   "annotations": [
     {
+      "timestamp": 1472470996199000,
+      "value": "cs",
       "endpoint": {
-        "serviceName": "zipkin-server",
-        "ipv4": "172.19.0.3",
-        "port": 9411
-      },
-      "timestamp": 1461750040359000,
-      "value": "cs"
+        "serviceName": "frontend",
+        "ipv4": "127.0.0.1"
+      }
     },
     {
+      "timestamp": 1472470996238000,
+      "value": "ws",
       "endpoint": {
-        "serviceName": "zipkin-server",
-        "ipv4": "172.19.0.3",
-        "port": 9411
-      },
-      "timestamp": 1461750040364000,
-      "value": "cr"
+        "serviceName": "frontend",
+        "ipv4": "127.0.0.1"
+      }
+    },
+    {
+      "timestamp": 1472470996403000,
+      "value": "wr",
+      "endpoint": {
+        "serviceName": "frontend",
+        "ipv4": "127.0.0.1"
+      }
+    },
+    {
+      "timestamp": 1472470996406000,
+      "value": "cr",
+      "endpoint": {
+        "serviceName": "frontend",
+        "ipv4": "127.0.0.1"
+      }
     }
   ],
   "binaryAnnotations": [
     {
-      "key": "jdbc.query",
-      "value": "select distinct `zipkin_spans`.`trace_id` from `zipkin_spans` join `zipkin_annotations` on (`zipkin_spans`.`trace_id` = `zipkin_annotations`.`trace_id` and `zipkin_spans`.`id` = `zipkin_annotations`.`span_id`) where (`zipkin_annotations`.`endpoint_service_name` = ? and `zipkin_spans`.`start_ts` between ? and ?) order by `zipkin_spans`.`start_ts` desc limit ?",
+      "key": "ca",
+      "value": true,
       "endpoint": {
-        "serviceName": "zipkin-server",
-        "ipv4": "172.19.0.3",
-        "port": 9411
+        "serviceName": "frontend",
+        "ipv4": "127.0.0.1",
+        "port": 49504
+      }
+    },
+    {
+      "key": "clnt/finagle.version",
+      "value": "6.36.0",
+      "endpoint": {
+        "serviceName": "frontend",
+        "ipv4": "127.0.0.1"
+      }
+    },
+    {
+      "key": "http.uri",
+      "value": "/api",
+      "endpoint": {
+        "serviceName": "frontend",
+        "ipv4": "127.0.0.1"
       }
     },
     {
       "key": "sa",
       "value": true,
       "endpoint": {
-        "serviceName": "mysql",
-        "ipv4": "172.19.0.2",
-        "port": 3306
+        "serviceName": "backend",
+        "ipv4": "127.0.0.1",
+        "port": 9000
       }
     }
-  ]
+  ],
+  "debug": false
 }