This includes cleanups and performance improvements when reading json.
Notably, this stops recreating lists by re-using Span.Builder

Before:

Benchmark                                               Mode  Cnt    Score    Error  Units
CodecBenchmarks.readClientSpan_json_zipkin              avgt   15    7.392 ±  0.087  us/op
CodecBenchmarks.readTenClientSpans_json_zipkin          avgt   15   59.769 ±  0.739  us/op
CodecBenchmarks.readTenTimesTenClientSpans_json_zipkin  avgt   15  608.130 ± 24.359  us/op

After:

Benchmark                                               Mode  Cnt    Score   Error  Units
CodecBenchmarks.readClientSpan_json_zipkin              avgt   15    7.337 ± 0.066  us/op
CodecBenchmarks.readTenClientSpans_json_zipkin          avgt   15   58.907 ± 0.671  us/op
CodecBenchmarks.readTenTimesTenClientSpans_json_zipkin  avgt   15  597.384 ± 7.373  us/op