@adriancole 2017-07-10
added a commit for thrift doing the same. will later run benchmarks there