@adriancole 2017-07-10
the "readTen" benchmarks are important to show optimizations when decoding lists which nowadays are often much larger than 10 spans. For example, most new tracers batch and can send quite a lot of spans per message (easily hundreds in some scenarios)